[AM][OPS-10184] connect legacy calculator to model with tests
